Implemented Facebook Login without the SDK. Here is the login flow:
You click the Facebook Login button and a request is send to the server to get the redirect_uri to the Facebook login page (the user is redirected when the redirect_uri is returned). Upon entering credentials and giving consent to the application, the user gets redirected back to the app with queryParams in the url containing a code which is then used to obtain an access_token from Facebook. This access_token is returned to the client and the client sends a request with it to the server to invoke the passport-facebook-token strategy and handle the user.